functionLayer/requirements.txt
line-bot-sdk
requests
PyGithub
fastapi[standard]